combine LensExample tests + add val check
